+++ database/schemas/Order.py
@@ -6,7 +6,7 @@ class Order(TimedBaseModel):
     __tablename__ = "orders"
     id = Column(BigInteger, primary_key=True, autoincrement=True)
     shop_id = Column(BigInteger, nullable=False)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     good_id = Column(Integer)
     good_name = Column(String)
     count = Column(Integer, default=1)
